+++ lib/src/ui/views/teams/magic_starter_team_create_view.dart
@@ -26,8 +26,7 @@ class _MagicStarterTeamCreateViewState
 
   @override
   void onInit() {
-    controller.clearErrors();
-    controller.setEmpty();
+    controller.resetQuietly();
   }
 
   @override